requirements.txt
beautifulsoup4==4.12.3
Flask==3.0.2
langchain==0.1.8
langchain_community==0.0.21
matplotlib==3.8.3
openai==1.12.0
pandas==2.2.0
python-dotenv==1.0.1
reportlab==4.0.9
Requests==2.31.0
sk==0.0.1
streamlit==1.31.1
duckduckgo-search==4.4.3
google-api-python-client==2.119.0